Change the DataType argument from Integer to Long in @Service, Repository

+++ src/main/java/com/ascendcorp/ctaworkshop/hotelchiangmai/repository/RoomRepository.java
@@ -21,7 +21,7 @@ public class RoomRepository {
         return query.getResultList();
     }
 
-    public Room findById(Integer id) {
+    public Room findById(Long id) {
         return entityManager.find(Room.class, id);
     }
